+++ src/main/java/uk/org/okapibarcode/backend/Code16k.java
@@ -765,7 +765,7 @@ public class Code16k extends Symbol {
                 } else {
                     black = true;
                 }
-                x += (double)(pattern[yBlock].charAt(xBlock) - '0');
+                x += pattern[yBlock].charAt(xBlock) - '0';
             }
             y += h;
             if (y > symbol_height) {
